% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut_test.R
\name{permut_test}
\alias{permut_test}
\title{Permutation test for a spatio- temporal point process on a linear network}
\usage{
permut_test(perm, Q_n, Xi, L0)
}
\arguments{
\item{perm}{Number of permutations to use}

\item{Q_n}{Data frame containing the spatio-temporal locations of the points to be permuted}

\item{Xi}{The point of the original point pattern}

\item{L0}{Linear network of the spatio-temporal point process to be permuted}
}
\value{
A vector of squared discrepancy values from the lista functions computed on the points of the pattern and their average
}
\description{
Permutation test for a spatio- temporal point process on a linear network
}
