% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ETASnet.R
\name{sim_ETASnet}
\alias{sim_ETASnet}
\title{Simulation of a spatio-temporal ETAS (Epidemic Type Aftershock Sequence) model on a linear network}
\usage{
sim_ETASnet(
  cat = NULL,
  params = c(0.078915, 0.003696, 0.013362, 1.2, 0.424466, 1.164793),
  betacov = 0.39,
  m0 = 2.5,
  b = 1.0789,
  tmin = 0,
  t.lag = 200,
  xmin = 600,
  xmax = 2200,
  ymin = 4000,
  ymax = 5300,
  iprint = TRUE,
  covdiag = FALSE,
  covsim = FALSE,
  L
)
}
\arguments{
\item{params}{A vector of parameters}

\item{tmin}{Minimum value of time.}

\item{iprint}{Default TRUE}

\item{covdiag}{Default FALSE}

\item{covsim}{Default FALSE}

\item{L}{linear network}
}
\value{
An ETAS catalogue on a linear network. It can be easily converted into an stlnpp object
}
\description{
This function simulates a spatio-temporal ETAS (Epidemic Type Aftershock Sequence) model on a linear network.
}
\examples{

\dontrun{
L0 = domain(chicago)
set.seed(95)
X0 = sim_ETASnet(cat = NULL, params = c(0.1293688525, 0.003696, 0.013362, 1.2,0.424466,  1.164793),
                         betacov = 0.5, m0 = 2.5, b = 1.0789, tmin = 0, t.lag = 200,
                         xmin = 600, xmax = 2200, ymin = 4000, ymax = 5300, iprint = TRUE,
                         covdiag = FALSE, covsim = FALSE, L = L0)$cat.sim
X <- as.stlpp(x = X0$long, y = X0$lat, t = X0$time,  L = L0)
X
plot(X)
}


}
