% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_LGCPnet.R
\name{sim_LGCPnet}
\alias{sim_LGCPnet}
\title{Simulation of a spatio-temporal LGCP (Log-Gaussian Cox Process) model on a linear network}
\usage{
sim_LGCPnet(
  lambda = lambda0,
  a = 0,
  b = 3,
  L = L0,
  check = FALSE,
  lmax = NULL,
  nsim = 1,
  sigma = 2,
  phi = 0.5,
  theta = 2,
  nmark = 1
)
}
\value{
an object of class stlpp if nsim=1, otherwise a list of objects of class stlpp
}
\description{
This function simulates a spatio-temporal LGCP (Log-Gaussian Cox Process) model on a linear network.
}
\examples{


\dontrun{
L0 = domain(chicago)
aa = 64 / ((1 - exp( - 4)) * (1 - exp( - 2)))
lambda0 = function(x, y, t) {exp(log(aa) - 4 * y - 2 * t)}
X = sim_LGCPnet(lambda0, a = 0, b = 4, nmark = 1, sigma = 2, phi = 0.5, theta = 1,L = L0, lmax = 10)
X
plot(X)
}


}
