% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STLginhom_local.R
\name{STLginhom_local}
\alias{STLginhom_local}
\title{Local inhomogeneous Spatio-temporal pair correlation functions on a linear network}
\usage{
STLginhom_local(X, lambda, normalize = FALSE, r = NULL, t = NULL, nxy = 10)
}
\arguments{
\item{X}{a realisation of a spatio-temporal point processes on a linear networks.}

\item{lambda}{values of estimated intensity.}

\item{normalize}{normalization factor to be considered.}

\item{r}{values of argument r where pair correlation function will be evaluated. optional.}

\item{t}{values of argument t where pair correlation function will be evaluated. optional.}

\item{nxy}{pixel array dimensions. optional.}
}
\value{
A list of objects of class sumstlpp.
}
\description{
This function calculates the local inhomogeneous pair correlation function for a spatio-temporal point processes on a linear network.
}
\examples{

\dontrun{
set.seed(10)
X <- rpoistlpp(.2, a = 0, b = 5, L = easynet)
X
lambda <- density(X, at = "points")
g <- STLginhom_local(X, lambda = lambda, normalize = TRUE)
#select an individual point
j = 1
g[[j]]
#plot the lista function and compare it with its theoretical value
inhom <- list(x = g[[j]]$r, y = g[[j]]$t, z = g[[j]]$ginhom)
theo <- list(x = g[[j]]$r, y = g[[j]]$t, z = g[[j]]$gtheo)
diff <- list(x = g[[j]]$r, y = g[[j]]$t, z = g[[j]]$ginhom - g[[j]]$gtheo)
par(mfrow=c(1,3))
fields::image.plot(inhom, main = "ginhom", col = hcl.colors(12, "YlOrRd", rev = FALSE), xlab = "Spatial distance", ylab = "Temporal distance")
fields::image.plot(theo, main = "gtheo", col = hcl.colors(12, "YlOrRd", rev = FALSE), xlab = "Spatial distance", ylab = "Temporal distance")
fields::image.plot(diff, main = "ginhom - gtheo", col = hcl.colors(12, "YlOrRd", rev = FALSE), xlab = "Spatial distance", ylab = "Temporal distance")
}




}
