% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STLginhom_local_a.R
\name{STLginhom_local_a}
\alias{STLginhom_local_a}
\title{Title}
\usage{
STLginhom_local_a(X, lambda, normalize = FALSE, r = NULL, t = NULL, nxy = 10)
}
\arguments{
\item{nxy}{}
}
\value{

}
\description{
Title
}
