% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STLKinhom_local_a.R
\name{STLKinhom_local_a}
\alias{STLKinhom_local_a}
\title{Title}
\usage{
STLKinhom_local_a(
  X,
  lambda = lambda,
  normalize = FALSE,
  r = NULL,
  t = NULL,
  nxy = 10
)
}
\arguments{
\item{nxy}{}
}
\value{

}
\description{
Title
}
