% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_local.R
\name{test_local}
\alias{test_local}
\title{Test of local structure for spatio-temporal point processes on linear networks}
\usage{
test_local(X, Z, method = "K", k, parallel = FALSE, ncores = 4)
}
\arguments{
\item{X}{Background spatio-temporal point pattern on a linear network. Must be an stlnpp object.}

\item{Z}{Other spatio-temporal point pattern on a linear network. Must be an stlnpp object.}

\item{method}{Character string indicating which version of LISTA function to use: either "K" or "g".
If "K", the Local spatio-temporal K-function on a linear network is used to run the test.
If "g", the Local spatio-temporal pair correlation function is used.}

\item{k}{Number of permutations}

\item{parallel}{FALSE is default. If TRUE, parallelization is performed.}

\item{ncores}{Number of parallel threads. Optional.}
}
\value{
A vector of p-values, one for each of the points in X.
}
\description{
Test of local structure for spatio-temporal point processes on linear networks
}
\examples{



\dontrun{
#simulate two spatio-temporal point patterns on a linear network with different local structure
L0 = domain(chicago)
set.seed(4)
X0 = sim_ETASnet(cat = NULL, params = c(0.078915/2, 0.003696,  0.013362,  1.2,0.424466,  1.164793),
                         betacov = 0.5, m0 = 2.5, b = 1.0789, tmin = 0, t.lag = 200,
                         xmin = 600, xmax = 2200, ymin = 4000, ymax = 5300,
                         iprint = TRUE, covdiag = FALSE, covsim = FALSE, L = L0)$cat.sim
X = as.stlpp(X0$long, X0$lat, X0$time, L = L0)
lambda0 <- 20 / (volume(domain(chicago)) * (200-25))
set.seed(4)
Z = rpoistlpp(lambda = lambda0, a = 25, b = 200, L = L0)

#compute the local test
set.seed(1); system.time(p <- test_local(X, Z, method = "K", k = 19, parallel = FALSE))

#display the significant and non-significan points
plot(X[which(p <= 0.05)])
plot(X[-which(p <= 0.05)])
}


}
